feat(backend): go服务接入trace close #2526
