+++ dbm-services/common/go-pubpkg/apm/trace/trace.go
@@ -17,14 +17,11 @@ import (
 	"dbm-services/common/go-pubpkg/logger"
 
 	"github.com/spf13/viper"
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/attribute"
-	oteltrace "go.opentelemetry.io/otel/trace"
 )
 
 const (
 	// TracerName trace name
-	TracerName      = "bk-dbm/db-apm"
+	TracerName      = "bk-dbm/http-tracer"
 	EnablePath      = "trace.enable"
 	HostPath        = "trace.otlp.host"
 	PortPath        = "trace.otlp.port"
@@ -42,88 +39,33 @@ var (
 	DataID                        int64
 )
 
-// InsertIntIntoSpan TODO
-func InsertIntIntoSpan(key string, value int, span oteltrace.Span) {
-	if span == nil {
-		return
-	}
-	span.SetAttributes(
-		attribute.Int(key, value),
-	)
-}
-
-// InsertStringIntoSpan TODO
-func InsertStringIntoSpan(key, value string, span oteltrace.Span) {
-	if span == nil {
-		return
-	}
-	span.SetAttributes(
-		attribute.String(key, value),
-	)
-}
-
-// InsertStringSliceIntoSpan TODO
-func InsertStringSliceIntoSpan(key string, value []string, span oteltrace.Span) {
-	if span == nil {
-		return
-	}
-	span.SetAttributes(
-		attribute.StringSlice(key, value),
-	)
-}
-
-// InsertIntSliceIntoSpan TODO
-func InsertIntSliceIntoSpan(key string, value []int, span oteltrace.Span) {
-	if span == nil {
-		return
-	}
-	span.SetAttributes(
-		attribute.IntSlice(key, value),
-	)
-}
-
-// IntoContext 填充trace，并返回处理后的context和span
-// span为nil时，说明没有开启trace
-func IntoContext(globalCtx context.Context, tracerName, spanName string) (context.Context, oteltrace.Span) {
-	var (
-		span     oteltrace.Span
-		traceCtx context.Context
-	)
-
-	// 向trace context中添加trace
-	tracer := otel.Tracer(tracerName)
-	traceCtx, span = tracer.Start(globalCtx, spanName)
-
-	return traceCtx, span
-
-}
-
 // Setup setupTrace
 func Setup() {
+	logger.Info(
+		"setup trace for service: %s, please check: \n"+
+			"TRACE_HOST: %s\n"+
+			"TRACE_PORT: %s\n"+
+			"TRACE_TOKEN: %s\n"+
+			"TRACE_TYPE: %s\n"+
+			"TRACE_SERVICE_NAME: %s\n"+
+			"TRACE_DATA_ID: %s\n",
+		ServiceName, otlpHost, otlpPort, otlpToken, OtlpType, DataID, ServiceName)
+
 	if OtlpEnable {
 		if otlpHost != "" && otlpPort != "" && otlpToken != "" && OtlpType != "" && DataID != 0 {
 			logger.Info("init trace for service: %s", ServiceName)
 			traceService := &Service{}
 			traceService.Reload(context.TODO())
 		} else {
-			logger.Error(
-				"init trace error for service: %s, please check: \n"+
-					"TRACE_HOST: %s\n"+
-					"TRACE_PORT: %s\n"+
-					"TRACE_TOKEN: %s\n"+
-					"TRACE_TYPE: %s\n"+
-					"TRACE_SERVICE_NAME: %s\n"+
-					"TRACE_DATA_ID: %s\n",
-				ServiceName, otlpHost, otlpPort, otlpToken, OtlpType, DataID, ServiceName)
+			logger.Error("skip init trace for service: %s, please check trace config: %s\n", ServiceName)
 		}
 	} else {
-		logger.Warn("skip init trace for service: %s", ServiceName)
+		logger.Warn("skip init trace for disabled service: %s", ServiceName)
 	}
 }
 
 // setDefaultConfig
 func setDefaultConfig() {
-	viper.SetEnvPrefix("APM")
 	viper.AutomaticEnv()
 
 	_ = viper.BindEnv(EnablePath, "TRACE_ENABLE")
@@ -149,7 +91,7 @@ func InitConfig() {
 	otlpHost = viper.GetString(HostPath)
 	otlpPort = viper.GetString(PortPath)
 	otlpToken = viper.GetString(TokenPath)
-	logger.Info("trace will Otlp to host->[%s] port->[%s] token->[%s]", otlpHost, otlpPort, otlpToken)
+	logger.Info("trace will Otlp url->[%s:%s] token->[%s] dataid->[%s]", otlpHost, otlpPort, otlpToken, DataID)
 
 	OtlpType = viper.GetString(OtlpTypePath)
 	logger.Info("trace will Otlp as %s type", OtlpType)
