+++ dbm-services/mysql/db-remote-service/pkg/apm/metric.go
@@ -1,3 +1,13 @@
+/*
+ * TencentBlueKing is pleased to support the open source community by making 蓝鲸智云-DB管理系统(BlueKing-BK-DBM) available.
+ * Copyright (C) 2017-2023 THL A29 Limited, a Tencent company. All rights reserved.
+ * Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at https://opensource.org/licenses/MIT
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
 package apm
 
 import "dbm-services/common/go-pubpkg/apm/metric"
@@ -7,6 +17,7 @@ var (
 	ExecuteCnt = "execute_cnt"
 )
 
+// CustomMetrics custom metric example
 var CustomMetrics = []*metric.Metric{
 	{
 		ID:          ErrCnt,
